chore(chromealive-ui): rename app timeline
