+++ apps/chromealive-ui/src/pages/timeline/views/FocusedTimeline.vue
