+++ apps/chromealive-ui/vue.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   outputDir: '../../build/apps/chromealive/ui',
   pages: {
-    app: './src/pages/app/index.ts',
+    timeline: './src/pages/timeline/index.ts',
     toolbar: './src/pages/toolbar/index.ts',
     databox: './src/pages/databox/index.ts',
     'pagestate-panel': './src/pages/pagestate-panel/index.ts',
